+++ lib/cocoapods-core/platform.rb
@@ -67,6 +67,14 @@ module Pod
       new :osx
     end
 
+    # Convenience method to initialize a macOS platform.
+    #
+    # @return [Platform] a macOS platform.
+    #
+    def self.macos
+      osx
+    end
+
     # Convenience method to initialize a tvOS platform.
     #
     # @return [Platform] a tvOS platform.
@@ -224,7 +232,7 @@ module Pod
     def self.string_name(symbolic_name)
       case symbolic_name
       when :ios then 'iOS'
-      when :osx then 'OS X'
+      when :osx then 'macOS'
       when :watchos then 'watchOS'
       when :tvos then 'tvOS'
       else symbolic_name.to_s