+++ lib/cocoapods-core/specification/dsl.rb
@@ -539,6 +539,7 @@ module Pod
       #
       def platform=(args)
         name, deployment_target = args
+        name = :osx if name.to_s == 'macos'
         attributes_hash['platforms'] = if name
                                          { name.to_s => deployment_target }
                                        else
@@ -1437,6 +1438,8 @@ module Pod
       def osx
         PlatformProxy.new(self, :osx)
       end
+
+      alias macos osx
 
       # Provides support for specifying tvOS attributes.
       #